separate stored pkm data from past gen lookups
