+++ src/app/App.tsx
@@ -16,9 +16,9 @@ import { AppInfoContext, appInfoInitialState, appInfoReducer, Settings } from '.
 import { DragMonContext, dragMonReducer } from '../state/dragMon'
 import { ErrorContext, errorReducer } from '../state/error'
 import { FilterContext, filterReducer } from '../state/filter'
-import { LookupContext, lookupReducer } from '../state/lookup'
 import { MouseContext, mouseReducer } from '../state/mouse'
 import { OpenSavesContext, openSavesReducer } from '../state/openSaves'
+import { PersistedPkmDataContext, persistedPkmDataReducer } from '../state/persistedPkmData'
 import { HomeData } from '../types/SAVTypes/HomeData'
 import './App.css'
 import AppTabs from './AppTabs'
@@ -77,7 +77,7 @@ function AppWithBackend() {
   const [mouseState, mouseDispatch] = useReducer(mouseReducer, { shift: false })
   const [dragMonState, dragMonDispatch] = useReducer(dragMonReducer, {})
   const [appInfoState, appInfoDispatch] = useReducer(appInfoReducer, appInfoInitialState)
-  const [lookupState, lookupDispatch] = useReducer(lookupReducer, { loaded: false })
+  const [lookupState, lookupDispatch] = useReducer(persistedPkmDataReducer, { loaded: false })
   const [filterState, filterDispatch] = useReducer(filterReducer, {})
   const [pluginState, pluginDispatch] = useReducer(pluginReducer, { plugins: [], loaded: false })
   const [settingsLoading, setSettingsLoading] = useState(false)
@@ -166,7 +166,7 @@ function AppWithBackend() {
     <PluginContext.Provider value={[pluginState, pluginDispatch]}>
       <AppInfoContext.Provider value={[appInfoState, appInfoDispatch, getEnabledSaveTypes]}>
         <MouseContext.Provider value={[mouseState, mouseDispatch]}>
-          <LookupContext.Provider value={[lookupState, lookupDispatch]}>
+          <PersistedPkmDataContext.Provider value={[lookupState, lookupDispatch]}>
             <OpenSavesContext.Provider
               value={[
                 openSavesState,
@@ -195,7 +195,7 @@ function AppWithBackend() {
                 </PokemonDragContextProvider>
               </DragMonContext.Provider>
             </OpenSavesContext.Provider>
-          </LookupContext.Provider>
+          </PersistedPkmDataContext.Provider>
         </MouseContext.Provider>
       </AppInfoContext.Provider>
     </PluginContext.Provider>
